+++ ReactiveCocoa/ObjC+Messages.swift
@@ -23,6 +23,9 @@ internal let NSMethodSignature: AnyClass = NSClassFromString("NSMethodSignature"
 
 // Methods of `NSInvocation`.
 @objc internal protocol ObjCInvocation {
+	@objc(setTarget:)
+	func setTarget(_ target: Any?)
+
 	@objc(setSelector:)
 	func objcSetSelector(_ selector: Selector)