+++ ReactiveCocoa/UIKit/iOS/UISearchBar.swift
@@ -2,45 +2,9 @@
 import ReactiveSwift
 import UIKit
 
-private class SearchBarDelegateProxy: DelegateProxy<UISearchBarDelegate>, UISearchBarDelegate {
-	@objc func searchBarTextDidBeginEditing(_ searchBar: UISearchBar) {
-		forwardee?.searchBarTextDidBeginEditing?(searchBar)
-	}
-
-	@objc func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
-		forwardee?.searchBarTextDidEndEditing?(searchBar)
-	}
-
-	@objc func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-		forwardee?.searchBar?(searchBar, textDidChange: searchText)
-	}
-	
-	@objc func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
-		forwardee?.searchBarCancelButtonClicked?(searchBar)
-	}
-
-	@objc func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-		forwardee?.searchBarSearchButtonClicked?(searchBar)
-	}
-
-	@objc func searchBarBookmarkButtonClicked(_ searchBar: UISearchBar) {
-		forwardee?.searchBarBookmarkButtonClicked?(searchBar)
-	}
-
-	@objc func searchBarResultsListButtonClicked(_ searchBar: UISearchBar) {
-		forwardee?.searchBarResultsListButtonClicked?(searchBar)
-	}
-
-	@objc func searchBar(_ searchBar: UISearchBar, selectedScopeButtonIndexDidChange selectedScope: Int) {
-		forwardee?.searchBar?(searchBar, selectedScopeButtonIndexDidChange: selectedScope)
-	}
-}
-
 extension Reactive where Base: UISearchBar {
-	private var proxy: SearchBarDelegateProxy {
-		return .proxy(for: base,
-		              setter: #selector(setter: base.delegate),
-		              getter: #selector(getter: base.delegate))
+	private var proxy: DelegateProxy<UISearchBarDelegate> {
+		return proxy(keyPath: \.delegate)
 	}
 
 	/// Sets the text of the search bar.
@@ -58,7 +22,8 @@ extension Reactive where Base: UISearchBar {
 	/// - note: To observe text values that change on all editing events,
 	///   see `continuousTextValues`.
 	public var textValues: Signal<String?, Never> {
-		return proxy.intercept(#selector(UISearchBarDelegate.searchBarTextDidEndEditing))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarTextDidEndEditing))
 			.map { [unowned base] in base.text }
 	}
 
@@ -66,44 +31,52 @@ extension Reactive where Base: UISearchBar {
 	///
 	/// - note: To observe text values only when editing ends, see `textValues`.
 	public var continuousTextValues: Signal<String?, Never> {
-		return proxy.intercept(#selector(proxy.searchBar(_:textDidChange:)))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBar(_:textDidChange:)))
 			.map { [unowned base] in base.text }
 	}
 
 	/// A signal of the latest selected scope button index upon any user selection.
 	public var selectedScopeButtonIndices: Signal<Int, Never> {
-		return proxy.intercept(#selector(proxy.searchBar(_:selectedScopeButtonIndexDidChange:)))
+		return proxy.reactive
+			.signal(for: #selector(proxy.delegateType.searchBar(_:selectedScopeButtonIndexDidChange:)))
 			.map { $0[1] as! Int }
 	}
 
 	/// A void signal emitted by the search bar upon any click on the cancel button
 	public var cancelButtonClicked: Signal<Void, Never> {
-		return proxy.intercept(#selector(proxy.searchBarCancelButtonClicked))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarCancelButtonClicked))
 	}
 
 	/// A void signal emitted by the search bar upon any click on the search button
 	public var searchButtonClicked: Signal<Void, Never> {
-		return proxy.intercept(#selector(proxy.searchBarSearchButtonClicked(_:)))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarSearchButtonClicked(_:)))
 	}
 
 	/// A void signal emitted by the search bar upon any click on the bookmark button
 	public var bookmarkButtonClicked: Signal<Void, Never> {
-		return proxy.intercept(#selector(proxy.searchBarBookmarkButtonClicked))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarBookmarkButtonClicked))
 	}
 
 	/// A void signal emitted by the search bar upon any click on the bookmark button
 	public var resultsListButtonClicked: Signal<Void, Never> {
-		return proxy.intercept(#selector(proxy.searchBarResultsListButtonClicked))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarResultsListButtonClicked))
 	}
 
 	/// A void signal emitted by the search bar upon start of editing
 	public var textDidBeginEditing: Signal<Void, Never> {
-		return proxy.intercept(#selector(proxy.searchBarTextDidBeginEditing))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarTextDidBeginEditing))
 	}
 
 	/// A void signal emitted by the search bar upon end of editing
 	public var textDidEndEditing: Signal<Void, Never> {
-		return proxy.intercept(#selector(proxy.searchBarTextDidEndEditing))
+		return proxy.reactive
+			.trigger(for: #selector(proxy.delegateType.searchBarTextDidEndEditing))
 	}
 
 	/// Shows and hides the cancel button of the search bar