+++ ReactiveCocoaTests/TestError.swift
@@ -15,8 +15,8 @@ internal extension SignalProducer {
 	/// Halts if an error is emitted in the receiver signal.
 	/// This is useful in tests to be able to just use `startWithNext`
 	/// in cases where we know that an error won't be emitted.
-	func assumeNoErrors() -> SignalProducer<Value, Never> {
-		return lift { $0.assumeNoErrors() }
+	func assumeNevers() -> SignalProducer<Value, Never> {
+		return lift { $0.assumeNevers() }
 	}
 }
 
@@ -24,7 +24,7 @@ internal extension Signal {
 	/// Halts if an error is emitted in the receiver signal.
 	/// This is useful in tests to be able to just use `startWithNext`
 	/// in cases where we know that an error won't be emitted.
-	func assumeNoErrors() -> Signal<Value, Never> {
+	func assumeNevers() -> Signal<Value, Never> {
 		return mapError { error in
 			fatalError("Unexpected error: \(error)")